requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
websockets>=12.0
pygame==2.5.2
SpeechRecognition==3.10.0
pyaudio==0.2.11
python-dotenv==1.0.0
supabase>=2.0.0
elevenlabs>=0.2.0
numpy>=1.21.0
soundfile>=0.12.0
psycopg2-binary>=2.9.0
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
Pillow>=9.0.0
svglib>=1.5.0
reportlab>=4.0.0
